requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124

torch
torchvision
kagglehub
Pillow
numpy
matplotlib
pandas
rasterio
scikit-image
scikit-learn